Mudando status de tarefa em collection
Former-commit-id: 3244fd8d1ccb1143e79ebf38289435fea497d053 [formerly 4bfe659fdf9d8bd03102a513ea8edc06a85e7635]
Former-commit-id: 78971ba4c6bfd95ff6581f3f7a6ec3b7ff0a821c

+++ src/br/com/lett/crawlernode/kernel/task/CrawlerSession.java
@@ -42,7 +42,7 @@ public class CrawlerSession {
 	private String url;
 
 	/** Processed id associated with the sku being crawled */
-	private int processedId;
+	private Long processedId;
 
 	/** Internal id associated with the sku being crawled */
 	private String internalId;
@@ -100,7 +100,7 @@ public class CrawlerSession {
 
 			// setting processed id
 			if (attrMap.containsKey(QueueService.PROCESSED_ID_MESSAGE_ATTR)) {
-				this.processedId = Integer.parseInt(attrMap.get(QueueService.PROCESSED_ID_MESSAGE_ATTR).getStringValue());
+				this.processedId = Long.parseLong(attrMap.get(QueueService.PROCESSED_ID_MESSAGE_ATTR).getStringValue());
 			}
 			
 			// setting session type
@@ -154,11 +154,11 @@ public class CrawlerSession {
 		this.sessionId = sessionId;
 	}
 
-	public int getProcessedId() {
+	public Long getProcessedId() {
 		return processedId;
 	}
 
-	public void setProcessedId(int processedId) {
+	public void setProcessedId(Long processedId) {
 		this.processedId = processedId;
 	}
